+++ ui/src/components/SSXChip/SSXChip.jsx
@@ -1,436 +1,54 @@
 /* eslint-disable react/jsx-key */
 
-/* eslint-disable jsx-a11y/control-has-associated-label */
-import 'fabric';
-import './ssxchipcontrol.css';
-
-import { useDispatch, useSelector } from 'react-redux';
 import { createSelector } from '@reduxjs/toolkit';
-import { executeCommand, setAttribute } from '../actions/beamline';
-
-
-import React, { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { Button, Card, Col, Container, Form, Row } from 'react-bootstrap';
-import { contextMenu, Item, Menu, Separator } from 'react-contexify';
+import { useDispatch, useSelector } from 'react-redux';
 
+import { executeCommand, setAttribute } from '../../actions/beamline';
+import { showTaskForm } from '../../actions/taskForm';
 import MotorInput from '../MotorInput/MotorInput';
+import { initChipCanvas, initFoilCanvas } from './ChipCanvasUtils';
+import ChipContextMenu from './ChipContextMenu';
+import styles from './ssxchipcontrol.module.css';
 
-const { fabric } = globalThis;
-
-const selectShapes = (state) => state.shapes;
-
-export const selectGrids = createSelector(
-  [selectShapes],
-  (shapes = {}) => {
-    const grids = {};
-
-    Object.values(shapes).forEach((shape) => {
-      if (shape.t === 'G') {
-        grids[shape.id] = shape;
-      }
-    });
-
-    return Object.values(grids);
-  }
-);
-
-function _GridData(fabricObject) {
-  return {
-    screenCoord: [fabricObject.top, fabricObject.left],
-    top: fabricObject.top,
-    left: fabricObject.left,
-    width: fabricObject.width,
-    height: fabricObject.height,
-    cellWidth: fabricObject.width,
-    cellHeight: fabricObject.height,
-    cellVSpace: 0,
-    cellHSpace: 0,
-    numCols: 1,
-    numRows: 1,
-    cellCountFun: null,
-    selected: false,
-    id: null,
-    result: null,
-    pixelsPerMMX: 1,
-    pixelsPerMMY: 1,
-  };
+function selectShapes(state) {
+  return state.shapes;
 }
 
-function ChipContextMenu({ onMoveTo, onAddTask }) {
-  return (
-    <Menu id="chip-context-menu">
-      <li aria-level="2" className="dropdown-header">
-        <b>Chip</b>
-      </li>
-      <Separator />
-      <Item id="moveto" data={{}} onClick={onMoveTo}>
-        Move to
-      </Item>
-      <Item id="addtask" data={{}} onClick={onAddTask}>
-        Add to queue
-      </Item>
-    </Menu>
-  );
-}
+const selectGrids = createSelector([selectShapes], (shapes = {}) => {
+  const grids = {};
 
-function showContextMenu(event, selection) {
-  contextMenu.show({
-    id: 'chip-context-menu',
-    event: event.e,
-    props: {
-      selection,
-    },
-  });
-}
-
-function renderChip(
-  chipSizeX,
-  chipSizeY,
-  rows,
-  cols,
-  blockSizeX,
-  blockSizeY,
-  spacing,
-  offset,
-  rowLabels = [],
-  colLabels = [],
-) {
-  const objects = [];
-
-  objects.push(
-    new fabric.Rect({
-      top: 0,
-      left: 0,
-      width: chipSizeX,
-      height: chipSizeY,
-      selectable: false,
-      hasControls: false,
-      borderColor: '#fff',
-      lockMovementX: true,
-      lockMovementY: true,
-      lockScalingX: true,
-      lockScalingY: true,
-      lockSkewingX: true,
-      lockSkewingY: true,
-      lockRotation: true,
-      hoverCursor: 'arrow',
-      type: 'CHIP',
-      objectIndex: [],
-    }),
-  );
-
-  // Add lables
-
-  for (let ci = 0; ci < cols; ci++) {
-    let label = (ci + 1).toString();
-
-    if (colLabels.length > 0) {
-      label = colLabels[ci];
-    }
-
-    objects.push(
-      new fabric.Text(label, {
-        top: offset / 2,
-        left:
-          ci * (blockSizeX + spacing) + offset + blockSizeX + blockSizeX / 4,
-        fontSize: blockSizeX * 0.7,
-        fontFamily: 'arial',
-        fill: '#f55',
-        objectCaching: false,
-        selectable: false,
-        hasControls: false,
-        borderColor: '#fff',
-        lockMovementX: true,
-        lockMovementY: true,
-        lockScalingX: true,
-        lockScalingY: true,
-        lockSkewingX: true,
-        lockSkewingY: true,
-        lockRotation: true,
-        hoverCursor: 'pointer',
-        type: 'BLOCK',
-      }),
-    );
-  }
-
-  for (let ri = 0; ri < rows; ri++) {
-    let label = (ri + 1).toString();
-
-    if (colLabels.length > 0) {
-      label = rowLabels[ri];
-    }
-
-    objects.push(
-      new fabric.Text(label, {
-        top: ri * (blockSizeY + spacing) + offset + blockSizeY,
-        left: offset / 2,
-        fontSize: blockSizeX * 0.7,
-        fontFamily: 'arial',
-        fill: '#f55',
-        objectCaching: false,
-        selectable: false,
-        hasControls: false,
-        borderColor: '#fff',
-        lockMovementX: true,
-        lockMovementY: true,
-        lockScalingX: true,
-        lockScalingY: true,
-        lockSkewingX: true,
-        lockSkewingY: true,
-        lockRotation: true,
-        hoverCursor: 'pointer',
-        type: 'BLOCK',
-      }),
-    );
-  }
-
-  for (let ri = 0; ri < rows; ri++) {
-    for (let ci = 0; ci < cols; ci++) {
-      objects.push(
-        new fabric.Rect({
-          top: ri * (blockSizeY + spacing) + offset + blockSizeY,
-          left: ci * (blockSizeX + spacing) + offset + blockSizeX,
-          width: blockSizeX,
-          height: blockSizeY,
-          fontFamily: 'arial',
-          fill: '#f55',
-          objectCaching: false,
-          hasControls: false,
-          borderColor: '#fff',
-          lockMovementX: true,
-          lockMovementY: true,
-          lockScalingX: true,
-          lockScalingY: true,
-          lockSkewingX: true,
-          lockSkewingY: true,
-          lockRotation: true,
-          hoverCursor: 'pointer',
-          type: 'BLOCK',
-          objectIndex: [ri, ci],
-        }),
-      );
-    }
-  }
-
-  return objects;
-}
-
-// function handleKeyDown(event) {
-//   if ([8, 46].includes(event.which)) {
-//     freeFormCanvas.remove(freeFormCanvas.getActiveObject());
-//     freeFormCanvas.renderAll();
-//   }
-
-//   return false;
-// }
-
-function initChipCanvas(currentChipLayout) {
-  const chipConfig = currentChipLayout.sections[0];
-
-  const numRows = chipConfig.number_of_rows;
-  const numCols = chipConfig.number_of_collumns;
-  const blockSizeX = chipConfig.block_size[0];
-  const blockSizeY = chipConfig.block_size[0];
-  const rowLabels = chipConfig.row_labels;
-  const colLabels = chipConfig.column_lables;
-
-  const offset = chipConfig.block_spacing[0];
-  const spacing = chipConfig.block_spacing[0];
-
-  const numTargetsX = chipConfig.targets_per_block[0];
-  const numTargetsY = chipConfig.targets_per_block[1];
-
-  const canvasWidth = numCols * (blockSizeX + spacing) + offset + blockSizeX;
-  const canvasHeight = numRows * (blockSizeY + spacing) + offset + blockSizeY;
-
-  const chipCanvas = new fabric.Canvas('chip-canvas', {
-    width: canvasWidth,
-    height: canvasHeight,
-    backgroundColor: '#CCC',
-    preserveObjectStacking: true,
-    altSelectionKey: 'ctrlKey',
-    selectionKey: 'ctrlKey',
-    fireRightClick: true,
-    stopContextMenu: true,
-    renderOnAddRemove: false,
-  });
-
-  const detailCanvas = new fabric.Canvas('chip-detail-canvas', {
-    width: canvasWidth,
-    height: canvasHeight,
-    backgroundColor: '#CCC',
-    preserveObjectStacking: true,
-    altSelectionKey: 'ctrlKey',
-    selectionKey: 'ctrlKey',
-    fireRightClick: true,
-    stopContextMenu: true,
-    renderOnAddRemove: false,
-  });
-
-  chipCanvas.on('mouse:down', (event) => {
-    const object = chipCanvas.findTarget(event.e);
-
-    if (event.button === 3) {
-      let selection = [];
-
-      if (object.type === 'BLOCK') {
-        selection.push(object.objectIndex);
-      }
-
-      if (object.type === 'activeSelection') {
-        selection = object._objects.map((o) => o.objectIndex);
-      }
-
-      if (selection.length > 0) {
-        chipCanvas.setActiveObject(object);
-        chipCanvas.requestRenderAll();
-        showContextMenu(event, selection);
-      }
+  Object.values(shapes).forEach((shape) => {
+    if (shape.t === 'G') {
+      grids[shape.id] = shape;
     }
   });
 
-  chipCanvas.on('selection:created', ({ selected, target }) => {
-    if (chipCanvas.getActiveObject()) {
-      chipCanvas.getActiveObject().lockMovementY = true;
-      chipCanvas.getActiveObject().lockMovementX = true;
-    }
-  });
+  return Object.values(grids);
+});
 
-  chipCanvas.on('selection:updated', ({ selected, target }) => {
-    if (chipCanvas.getActiveObject()) {
-      chipCanvas.getActiveObject().lockMovementY = true;
-      chipCanvas.getActiveObject().lockMovementX = true;
-    }
-  });
-
-  chipCanvas.add(
-    ...renderChip(
-      canvasWidth,
-      canvasHeight,
-      numRows,
-      numCols,
-      blockSizeX,
-      blockSizeY,
-      spacing,
-      offset,
-      rowLabels,
-      colLabels,
-    ),
-  );
-
-  chipCanvas.requestRenderAll();
-
-  detailCanvas.add(
-    ...renderChip(
-      canvasWidth,
-      canvasHeight,
-      numTargetsX,
-      numTargetsX,
-      canvasWidth / numTargetsX - (spacing / numTargetsX) * 4,
-      canvasHeight / numTargetsY - (spacing / numTargetsY) * 4,
-      spacing / numTargetsX,
-      offset,
-    ),
-  );
-  detailCanvas.renderAll();
-
-  return [chipCanvas, detailCanvas];
-}
-
-function initFoilCanvas(gridList) {
-  const freeFormCanvas = new fabric.Canvas('chip-free-form-canvas', {
-    width: 300,
-    height: 300,
-    backgroundColor: '#CCC',
-    preserveObjectStacking: true,
-    altSelectionKey: 'ctrlKey',
-    selectionKey: 'ctrlKey',
-    fireRightClick: true,
-    stopContextMenu: true,
-    renderOnAddRemove: false,
-  });
-
-  freeFormCanvas.on('mouse:down', (event) => {
-    const pointer = freeFormCanvas.getPointer(event.e);
-
-    if (!event.e.altKey) {
-      return;
-    }
-    freeFormCanvas.discardActiveObject();
-
-    const rect = new fabric.Rect({
-      left: pointer.x,
-      top: pointer.y,
-      originX: 'left',
-      originY: 'top',
-      width: 0,
-      height: 0,
-      angle: 0,
-      fill: 'rgba(255,0,0,0.5)',
-      transparentCorners: false,
-    });
-
-    freeFormCanvas.add(rect);
-    freeFormCanvas.setActiveObject(rect);
-    freeFormCanvas.renderAll();
-  });
-
-  freeFormCanvas.on('mouse:move', (event) => {
-    if (!event.e.altKey) {
-      return;
-    }
-
-    const mouse = freeFormCanvas.getPointer(event);
-    const rect = freeFormCanvas.getActiveObject();
-
-    const w = Math.abs(mouse.x - rect.left);
-    const h = Math.abs(mouse.y - rect.top);
-
-    if (!w || !h) {
-      return;
-    }
-
-    rect.set('width', w).set('height', h);
-
-    freeFormCanvas.renderAll();
-  });
-
-  // freeFormCanvas.on('mouse:up', (evnt) => {
-  //   handleAddGrid(_GridData(freeFormCanvas.getActiveObject()));
-  //   freeFormCanvas.discardActiveObject();
-  //   freeFormCanvas.renderAll();
-  // });
-
-  gridList.forEach((gridData) => {
-    freeFormCanvas.add(
-      new fabric.Rect({
-        left: gridData.screenCoord[1],
-        top: gridData.screenCoord[0],
-        originX: 'left',
-        originY: 'top',
-        width: gridData.width,
-        height: gridData.height,
-        angle: 0,
-        fill: 'rgba(255,0,0,0.5)',
-        transparentCorners: false,
-      }),
-    );
-  });
-
-  freeFormCanvas.renderAll();
-
-  return freeFormCanvas;
-}
-
-export default function SSXChip(props) {
+export default function SSXChip() {
   const canvasRef = useRef(null);
   const detailCanvasRef = useRef(null);
   const freeFormCanvasRef = useRef(null);
 
   const dispatch = useDispatch();
+  const currentSampleID = useSelector((state) => state.queue.currentSampleID);
+  const sampleData = useSelector(
+    (state) => state.sampleGrid.sampleList[currentSampleID],
+  );
+  const defaultParameters = useSelector(
+    (state) => state.taskForm.defaultParameters,
+  );
+  const groupFolder = useSelector((state) => state.queue.groupFolder);
   const gridList = useSelector(selectGrids);
-  const hardwareObjects = useSelector((state) => state.beamline.hardwareObjects);
-  const uiproperties = useSelector((state) => state.uiproperties.sample_view_motors);
+  const hardwareObjects = useSelector(
+    (state) => state.beamline.hardwareObjects,
+  );
+  const uiproperties = useSelector(
+    (state) => state.uiproperties.sample_view_motors,
+  );
 
   const sampleVerticalUiProp = uiproperties.components.find(
     (el) => el.role === 'sample_vertical',
@@ -442,37 +60,32 @@ export default function SSXChip(props) {
 
   const focus = uiproperties.components.find((el) => el.role === 'focus');
 
-  const sampleMotorVertical =
-    hardwareObjects[sampleVerticalUiProp.attribute];
+  const sampleMotorVertical = hardwareObjects[sampleVerticalUiProp.attribute];
   const sampleMotorHorizontal =
     hardwareObjects[sampleHorizontalUiProp.attribute];
   const focusMotor = hardwareObjects[focus.attribute];
-
 
   const headConfiguration =
     hardwareObjects.diffractometer.attributes.head_configuration ?? {};
 
   const chipLayoutList = headConfiguration.available;
-  const currentChipLayout = chipLayoutList[headConfiguration.current];
-
+  const _chipLayout = chipLayoutList[headConfiguration.current];
 
   const [chipState, setChipState] = useState({
-    top_left_x: currentChipLayout.calibration_data.top_left[0],
-    top_left_y: currentChipLayout.calibration_data.top_left[1],
-    top_left_z: currentChipLayout.calibration_data.top_left[2],
-    top_right_x: currentChipLayout.calibration_data.top_right[0],
-    top_right_y: currentChipLayout.calibration_data.top_right[1],
-    top_right_z: currentChipLayout.calibration_data.top_right[2],
-    bottom_left_x: currentChipLayout.calibration_data.bottom_left[0],
-    bottom_left_y: currentChipLayout.calibration_data.bottom_left[1],
-    bottom_left_z: currentChipLayout.calibration_data.bottom_left[2],
-    bottom_right_x: currentChipLayout.calibration_data.bottom_right[0],
-    bottom_right_y: currentChipLayout.calibration_data.bottom_right[1],
-    bottom_right_z: currentChipLayout.calibration_data.bottom_right[2],
-    currentLayoutName: currentLayoutName,
+    top_left_x: _chipLayout.calibration_data.top_left[0],
+    top_left_y: _chipLayout.calibration_data.top_left[1],
+    top_left_z: _chipLayout.calibration_data.top_left[2],
+    top_right_x: _chipLayout.calibration_data.top_right[0],
+    top_right_y: _chipLayout.calibration_data.top_right[1],
+    top_right_z: _chipLayout.calibration_data.top_right[2],
+    bottom_left_x: _chipLayout.calibration_data.bottom_left[0],
+    bottom_left_y: _chipLayout.calibration_data.bottom_left[1],
+    bottom_left_z: _chipLayout.calibration_data.bottom_left[2],
+    bottom_right_x: _chipLayout.calibration_data.bottom_right[0],
+    bottom_right_y: _chipLayout.calibration_data.bottom_right[1],
+    bottom_right_z: _chipLayout.calibration_data.bottom_right[2],
+    currentLayoutName: headConfiguration.current,
   });
-
-  const [_canvas, setCanvas] = useState('');
 
   const [activePosition, setActivePosition] = useState('');
 
@@ -497,7 +110,6 @@ export default function SSXChip(props) {
       ],
     }));
   }
-
 
   function handleAddTask(triggerEvent) {
     const sid = -1;
@@ -524,17 +136,12 @@ export default function SSXChip(props) {
     );
   }
 
-  // function handleAddGrid(data) {
-  //   dispatch(addShape({ t: 'G', ...data }));
-  // }
-
   function initCanvas() {
-    const currentChipLayout = chipLayoutList[currentLayoutName];
-    const holderType = currentChipLayout.holder_type;
+    const holderType = chipLayoutList[chipState.currentLayoutName].holder_type;
     let canvas = null;
 
     if (holderType === 'KNOWN_GEOMETRY') {
-      canvas = initChipCanvas(currentChipLayout);
+      canvas = initChipCanvas(chipLayoutList[chipState.currentLayoutName]);
     } else if (holderType === 'FREE_GEOMETRY') {
       canvas = initFoilCanvas(gridList);
     }
@@ -543,28 +150,17 @@ export default function SSXChip(props) {
   }
 
   useEffect(() => {
-    setCanvas(initCanvas());
+    initCanvas();
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [chipState]);
 
-  function handleInputValueChange(key, event) {
-    setChipState((prevState) => ({
-      ...prevState,
-      [key]: event.target.value,
-    }));
-  }
-
-  function handleSubmit(key, arg, e) {
+  function handleSubmit(key, e, arg) {
     switch (key) {
       case 'move_to': {
-        dispatch(setAttribute(
-          sampleMotorHorizontal.name,
-          chipState[`${arg}_x`],
-        ));
-        dispatch(setAttribute(
-          sampleMotorVertical.name,
-          chipState[`${arg}_y`],
-        ));
+        dispatch(
+          setAttribute(sampleMotorHorizontal.name, chipState[`${arg}_x`]),
+        );
+        dispatch(setAttribute(sampleMotorVertical.name, chipState[`${arg}_y`]));
         dispatch(setAttribute(focusMotor.name, chipState[`${arg}_z`]));
         break;
       }
@@ -603,40 +199,55 @@ export default function SSXChip(props) {
         break;
       }
       case 'set_layout': {
-        const currentChipLayout = chipLayoutList[e.target.value];
+        const chipLayout = chipLayoutList[e.target.value];
 
         setChipState({
-          top_left_x: currentChipLayout.calibration_data.top_left[0],
-          top_left_y:
-            currentChipLayout.calibration_data.todiffractometer.focusp_left[1],
-          top_left_z: currentChipLayout.calibration_data.top_left[1],
-          top_right_x: currentChipLayout.calibration_data.top_right[0],
-          top_right_y: currentChipLayout.calibration_data.top_right[0],
-          top_right_z: currentChipLayout.calibration_data.top_right[1],
-          bottom_left_x: currentChipLayout.calibration_data.bottom_left[0],
-          bottom_left_y: currentChipLayout.calibration_data.bottom_left[0],
-          bottom_left_z: currentChipLayout.calibration_data.bottom_left[1],
+          top_left_x: chipLayout.calibration_data.top_left[0],
+          top_left_y: chipLayout.calibration_data.top_left[1],
+          top_left_z: chipLayout.calibration_data.top_left[2],
+          top_right_x: chipLayout.calibration_data.top_right[0],
+          top_right_y: chipLayout.calibration_data.top_right[1],
+          top_right_z: chipLayout.calibration_data.top_right[2],
+          bottom_left_x: chipLayout.calibration_data.bottom_left[0],
+          bottom_left_y: chipLayout.calibration_data.bottom_left[1],
+          bottom_left_z: chipLayout.calibration_data.bottom_left[2],
           currentLayoutName: e.target.value,
         });
 
-        dispatch(executeCommand('diffractometer', 'diffractometer', 'set_chip_layout', {
-          layout_name: e.target.value,
-        }));
+        dispatch(
+          executeCommand(
+            'diffractometer',
+            'diffractometer',
+            'set_chip_layout',
+            {
+              layout_name: e.target.value,
+            },
+          ),
+        );
 
         break;
       }
       case 'apply': {
-        dispatch(executeCommand(
-          'diffractometer',
-          'diffractometer',
-          'use_position_for_calibration',
-          { data: calibratedPositions },
-        ));
+        dispatch(
+          executeCommand(
+            'diffractometer',
+            'diffractometer',
+            'use_position_for_calibration',
+            { data: calibratedPositions },
+          ),
+        );
 
         break;
       }
       case 'ir_auto_focus': {
-        dispatch(executeCommand('diffractometer', 'diffractometer', 'ir_auto_focus', {}));
+        dispatch(
+          executeCommand(
+            'diffractometer',
+            'diffractometer',
+            'ir_auto_focus',
+            {},
+          ),
+        );
 
         break;
       }
@@ -652,7 +263,7 @@ export default function SSXChip(props) {
     return activePosition === name ? 'outline-primary' : 'outline-secondary';
   }
 
-  const holderType = currentChipLayout.holder_type;
+  const holderType = chipLayoutList[chipState.currentLayoutName].holder_type;
   const chipVisible = holderType === 'KNOWN_GEOMETRY' ? '' : 'd-none';
   const foilVisible = holderType === 'FREE_GEOMETRY' ? '' : 'd-none';
   const calibratedCount = Object.values(calibratedPositions).filter(
@@ -662,7 +273,7 @@ export default function SSXChip(props) {
   return (
     <Container className="chip-container">
       <Row>
-        <Col xs={8}>
+        <Col xs={12}>
           <Card>
             <Card.Body>
               <div>
@@ -679,11 +290,13 @@ export default function SSXChip(props) {
                           }
                           value={chipState.currentLayoutName}
                         >
-                          {headConfiguration.available.map((item) => (
-                            <option key={item} value={item}>
-                              {item}
-                            </option>
-                          ))}
+                          {Object.keys(headConfiguration.available).map(
+                            (item) => (
+                              <option key={item} value={item}>
+                                {item}
+                              </option>
+                            ),
+                          )}
                         </Form.Select>
                       </Col>
                     </Form.Group>
@@ -691,19 +304,29 @@ export default function SSXChip(props) {
                 </Row>
               </div>
               <div className={chipVisible}>
-                <div className="chip-canvas-container">
-                  <canvas id="chip-canvas" ref={canvasRef} />
-                  <ChipContextMenu
-                    onAddTask={handleAddTask}
+                <div className={styles.chipCanvasContainer}>
+                  <canvas
+                    aria-label="chip-canvas"
+                    id="chip-canvas"
+                    ref={canvasRef}
                   />
+                  <ChipContextMenu onAddTask={handleAddTask} />
                 </div>
-                <div className="chip-detial-canvas-container">
-                  <canvas id="chip-detail-canvas" ref={detailCanvasRef} />
+                <div className={styles.chipCanvasContainer}>
+                  <canvas
+                    aria-label="chip-detials-canvas"
+                    id="chip-detail-canvas"
+                    ref={detailCanvasRef}
+                  />
                 </div>
               </div>
               <div className={foilVisible}>
                 <div className="chip-free-form-canvas-container">
-                  <canvas id="chip-free-form-canvas" ref={freeFormCanvasRef} />
+                  <canvas
+                    aria-label="chip-free-form-canvas"
+                    id="chip-free-form-canvas"
+                    ref={freeFormCanvasRef}
+                  />
                 </div>
               </div>
               <Row style={{ 'padding-top': 10 }}>
@@ -717,7 +340,7 @@ export default function SSXChip(props) {
                           'text-transform': 'capitalize',
                           'margin-right': 10,
                         }}
-                        onClick={(e) => handleSubmit('move_to', name, e)}
+                        onClick={(e) => handleSubmit('move_to', e, name)}
                       >
                         {name.replace('_', ' ')}
                       </Button>
@@ -743,9 +366,12 @@ export default function SSXChip(props) {
         </Col>
       </Row>
       <Row>
-        <Col xs={8}>
+        <Col xs={12}>
           <Card>
             <Card.Body>
+              <Row>
+                <h5>Chip calibratation:</h5>
+              </Row>
               <Row>
                 <Col className="col-sm-auto pe-0">
                   <Button
