--- ui/src/components/SSXChip/ssxchipcontrol.css
@@ -1,18 +1,0 @@
-#test {
-  min-width: fit-content;
-}
-
-.chip-title {
-  font-weight: bold;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.chip-detial-canvas-container {
-  display: inline-block;
-}
-
-.chip-canvas-container {
-  display: inline-block;
-}
