+++ ui/src/components/SSXChip/ssxchipcontrol.module.css
@@ -1,0 +1,3 @@
+.chipCanvasContainer {
+  display: inline-block;
+}
