+++ ui/src/containers/SampleViewContainer.jsx
@@ -1,9 +1,6 @@
 import { Col, Container, Row } from 'react-bootstrap';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
-import { executeCommand, setAttribute } from '../actions/beamline';
-import { addShape } from '../actions/sampleview';
-import { showTaskForm } from '../actions/taskForm';
 import PlateManipulator from '../components/Equipment/PlateManipulator';
 import motorInputStyles from '../components/MotorInput/MotorInput.module.css';
 import ApertureInput from '../components/SampleView/ApertureInput';
@@ -11,7 +8,6 @@ import ContextMenu from '../components/SampleView/ContextMenu';
 import MotorControls from '../components/SampleView/MotorControls';
 import PhaseInput from '../components/SampleView/PhaseInput';
 import SampleImage from '../components/SampleView/SampleImage';
-import SSXChipControl from '../components/SSXChip/SSXChipControl';
 import BeamlineSetupContainer from './BeamlineSetupContainer';
 import DefaultErrorBoundary from './DefaultErrorBoundary';
 import SampleQueueContainer from './SampleQueueContainer';
@@ -24,20 +20,8 @@ function getShapes(shapes, type) {
 }
 
 function SampleViewContainer() {
-  const dispatch = useDispatch();
-
-  const sampleList = useSelector((state) => state.sampleGrid.sampleList);
-  const currentSampleID = useSelector((state) => state.queue.currentSampleID);
-  const groupFolder = useSelector((state) => state.queue.groupFolder);
-  const hardwareObjects = useSelector(
-    (state) => state.beamline.hardwareObjects,
-  );
-  const defaultParameters = useSelector(
-    (state) => state.taskForm.defaultParameters,
-  );
   const shapes = useSelector((state) => state.shapes.shapes) || {};
   const uiproperties = useSelector((state) => state.uiproperties);
-  const mode = useSelector((state) => state.general.mode);
   const sampleChangerContents = useSelector(
     (state) => state.sampleChanger.contents,
   );
@@ -97,25 +81,6 @@ function SampleViewContainer() {
                 </label>
                 <ApertureInput />
               </div>
-            )}
-
-            {mode === 'SSX-CHIP' && (
-              <SSXChipControl
-                showForm={(...args) => dispatch(showTaskForm(...args))}
-                currentSampleID={currentSampleID}
-                sampleData={sampleList[currentSampleID]}
-                defaultParameters={defaultParameters}
-                groupFolder={groupFolder}
-                hardwareObjects={hardwareObjects}
-                uiproperties={uiproperties.sample_view_motors}
-                addShape={(...args) => dispatch(addShape(...args))}
-                grids={grids}
-                selectedGrids={selectedGrids}
-                setAttribute={(...args) => dispatch(setAttribute(...args))}
-                sendExecuteCommand={(...args) =>
-                  dispatch(executeCommand(...args))
-                }
-              />
             )}
             {sampleChangerContents.name === 'PlateManipulator' && (
               <PlateManipulator inPopover />
